% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixvectormultiply.R
\name{matrixvectormultiply}
\alias{matrixvectormultiply}
\title{Compute matrix-vector product ABx}
\usage{
matrixvectormultiply(A, B, x, switch_var)
}
\arguments{
\item{A}{A matrix}

\item{B}{A matrix}

\item{x}{A column vector}

\item{switch_var}{A boolean}
}
\description{
Compute matrix-vector product ABx
}
